fix: Removes passing and init browser for TRF5 as it is not used
We were passing and trying to start a headless browser but since we are not really using it, it usually failed fetching documents for not being able to start a new session. Removing the browser startup aims to accelerate the crawler and prevent uneeded errors

Some indentations fixes, but we still need to standardize the project.

Signed-off-by: Ricardo Verhaeg <ricardo.verhaeg@inspira.legal>